refactor: re-structured code base into service crates
