+++ src/service/mod.rs
@@ -1,0 +1,4 @@
+pub mod config;
+pub mod message;
+pub mod discord;
+pub mod timer;
